[IMP] documents: send document to trash when deleting the record
Purpose
=======
When we delete a record, all the related attachments are
deleted and so, the related documents too.

This might delete some documents when the user doesn't expect it,
so now we send the documents to the trash instead.

Technical
=========
This could have been implemented by overwriting `BaseModel` like
we did for standard attachments, but to not impact the performance
of the unlink method of all Odoo models, it has been done with a mixin,
and we choose for which models we want that behavior.

Task-3777950

closes odoo/enterprise#61839

Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>