[FIX] stock_barcode: retrieve right record in cache
Steps to reproduce:
- Install `stock_barcode`
- Create two product "aaa" and "bbb" with both of them being tracked by
  lots with the same lot number "123"
- Open the Inventory app and go to Configuration > Operation Types >
  Delivery Orders
- Enable Product in the Barcode App
- Go to the barcode app > Operations > Delivery Orders and click on New
- Scan 123
- Scan aaa (or bbb it probably depends on order of creation of product)
- Scan 123

Issues:
bbb has been added even though we never scanned it.

The reason for this bug is that when we scan 123 we don't have it in the
cache as such we make an RPC and store 123 in `missCache`.

https://github.com/odoo/enterprise/blob/ed30bde4af5d77261f3a44ba4235fd435153c7ce/stock_barcode/static/src/lazy_barcode_cache.js#L184-L185

After this when we scan aaa we retrieve the record, and add the line as
intended. The problem arise when we scan 123, since it's already in our
cache.

https://github.com/odoo/enterprise/blob/ed30bde4af5d77261f3a44ba4235fd435153c7ce/stock_barcode/static/src/lazy_barcode_cache.js#L130

However that record is not right since this lot is linked to the wrong
product bbb.

https://github.com/odoo/enterprise/blob/ed30bde4af5d77261f3a44ba4235fd435153c7ce/stock_barcode/static/src/lazy_barcode_cache.js#L135-L138

We will try to retrieve again the right record, however we will once
again get a hit on the cache. As such no RPC call will be made to
retrieve the lot we're looking for.

We will go in the function once again, where we will have the same
problem this time in the `if (model)` condition.

https://github.com/odoo/enterprise/blob/ed30bde4af5d77261f3a44ba4235fd435153c7ce/stock_barcode/static/src/lazy_barcode_cache.js#L112-L115

Later in the execution we will get to this line where we call the
function without a filters which will allow the wrong record to be
added.

https://github.com/odoo/enterprise/blob/ed30bde4af5d77261f3a44ba4235fd435153c7ce/stock_barcode/static/src/models/barcode_model.js#L943

The root of this problem lies in the way we set the key to the cache,
since two product can share the same barcode we need a more precise
cache key to differentiate them.

The proposed solutions is to cache result based on the provided
arguments to the function. This means fewer hit if we retrieve the same
record with different filter for example, however it's safer as when we
hit we are sure to get the right record.

opw-3862263

closes odoo/enterprise#61844

Signed-off-by: Mattis Megevand (mmeg) <mmeg@odoo.com>