[REF] mrp_mps: forecast details wizard buttons - visual changes
The statinfo widget shows the quantities followed by the text.
This is confusing because when it shows "120 Receipts", it means
there's a receipt with a quantity of 120 and not 120 actual receipts.

With this PR, we remove the widget and modify the color of the button
depending on if there's actual quantities or not.

Part-of: odoo/enterprise#56128