[FIX] mrp_mps: correct calculate with lead times
You have a 3 products linked by 2 boms:
- final product, manufactured with 1 semifin, lead time = 1 day
- semifin, manufactured with 1 component, lead time = 1 day
- component, no bom

You add the final product while precising the bom, it will add the
semifin and the component.
You put 1 unit in the Forecasted Demand of the final product in July:
- semifin will have an indirect demand of 1 for June: OK
- component will have an indirect demand of 1 for May: not OK

This is because the MPS calculates the indirect demand date based on
the date_start of the period of demand of its parent product.
So if you need 1 unit of final product in July, it uses the 1st of
October minus the lead time of 1 day => 30th of June for the semifin
Because it is in June, it considers that it needs 1 semifin at the
start of June, so 1st of June minus the lead time of 1 day
=> 31st of May for the component => period of May.

This commit stores the related_date in a dictionary with the
period (date_start, date_stop) plus the product as a key.
It then uses this related_date when calculating the related_date of its
children. So in the case above the indirect demand of component will be
calculated based on the related_date of semifin (30th of June) instead
of the date_start of its period. Thus the final date will be 29th of June
instead of 31st of May.

Part-of: odoo/enterprise#56128