[IMP] sign: add default sign template for activity
Purpose
=======
When user schedule activity of 'Request Signature' type, and request for
signature then needs to select a template for signature. If user has important
role in company then need to browse through lot of templates and if not then
don't know which template to be selected.

After this commit
=================
When user create activity type with 'Request Signature' action then user will be
able to select default template over there and it will populate as default
template to scheduled activity of 'Request Signature' type. Also user can change
the template while requesting the signature.

Task-3895712

closes odoo/enterprise#62032

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>