[IMP] l10n_be_hr_payroll_fleet: Change max unused cars default value in fleet settings, adjust available cars computation.
A warning in the salary cofigurator is displayed
if the number of available cars in a company is
greater than the specified number in the fleet
app settings (`fleet app -> configuration ->
settings -> 'New Vehicle Request'`).

The changes target the default value of the
'New Vehicle Request' setting and set it to 5
instead of 1000 because a company would unlikely
have 1000 available cars.

Also, add a tooltip to `New Vehicle Request` setting in
the fleet app to explain the setting more.

In addition to that, the changes target the computation of the
available cars. Any car in a stage that is
marked as 'hide_in_offer' will not be counted as
available (The different stages are in the fleet
app. You can edit a stage to mark it as
'hide_in_offer')

task-3530888

closes odoo/enterprise#62051

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Co-authored-by: Tivisse <yti@odoo.com>