[REF] mrp_mps: add route and vendor to MPS
This commit needs PR odoo/odoo#159423 to work.
We want to allow the user to set a specific route + vendor (if needed)
on each mps record. This needs to follow the same logic as the replenishment
wizard found on the product form, i.e. the `allowed_route_ids` logic.

For that, part of the code of the replenishment wizard was turned into
a mixin. This avoids duplicating code and creating multiple submodules
for the overrides in 'Drop shipping', 'Dropship and Subcontracting Management'
and 'Dropship and Subcontracting Management'.

Part-of: odoo/enterprise#56128