[IMP] sign: improve sign request wizard
before this commit, user were encountering confusion regarding the company field
leading to the necessity of manually setting this field, the current user
wasn't automatically assigned to the company field.

In this commit, the company field has been relabeled as user for clarity.
Additionally, the default value for the user field is now automatically set
when a user is set.

task-3890179

closes odoo/enterprise#62172

Related: odoo/upgrade#6121
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>