[IMP] mrp_mps: replenishment trigger
We want to be able to run a cron to automatically create replenishment
for specific MPS records. We also want to set some MPS recors as to
never be replenished from the MPS view.

replenish_trigger adds that behavior with 3 states:
- manual: this is how the MPS currently works and the default value
- automated: records with that setting will trigger action_replenish in the cron
- never: records with that setting cannot trigger a replenishment from the MPS view

Additionally, this allows to hide the 'Order' button from the 'Replenishment' row.

Part-of: odoo/enterprise#56128