[REF] mrp_mps: set_forecast_qty
Currently when changing quantities, Odoo will look for the first
forecast it finds in the corresponding period and add the difference
on it. If no forecast is found, it will create a new one at the end
of the period.

This commit changes the way quantities are added/removed from
Forecasted Demand.
Adding quantity:
- first it will search for the first forecast of the period
- if the first forecast has the same date as date_start, it will add the quantity there
- else it will create a new forecast at the start of the period

Removing quantity:
- first it will loop through all forecasts of the period in reverse date order
- it will then try to substract the difference on each forecast, starting with the latest
- each forecast will be set to zero until the difference has been spent
- if all forecasts have been looped through and there's still a difference, the first forecast will receive a negative quantity
- if not, the current forecast in the loop will see its quantity reduced by what's left of the difference

Part-of: odoo/enterprise#56128