[FIX] account_asset: extend length with increase in linear
Create an asset for 2400$, acquisition date on 1/1/2023.
Reevaluate it on 31/12/2023, 36 montths (+12) and add 1200$ to the amount.

=> The entries on the parent asset finish way too early.
You don't get 36 months of lifetime.
We would expect the amount of each new depreciation to be 50,
so the amount of the parent + the child equals 100 each month for 36 months

The reason why it happened is that we tried to compute the expected amount
after each depreciation. This took into account changes of amount, but not
changes of length.
We don't have a good way to incorporate changes of length in the old computation.
To fix this, for linear, we consider the original amount
from the beginning/the reevaluation to compute the amount for each period.

It does not hold for degressive_then_linear, so we separate the logic.

opw-3811165

closes odoo/enterprise#63247

X-original-commit: e7b9a727ceab4697ba6a7a1822504f690edefae3
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>