[FIX] l10n_mx_reports: Fix parsing of SAT XML code
The field NumCta in the SAT XML is parsed from the lines in the trial
balance report. The name of the line is the code for the account
concatenated with the name of the account. The existing
implementation assumes the code has 3 sets of digits separated by two
dots, and that the account name has no dots. This is true for the
default chart of accounts for Mexico, but is not necessarily the case
if additional accounts are added by the user.

As a result, invalid NumCta values were generated, which would
include parts of the account name if dots were present in it,
or the value would be too long if the code had more than two dots.

Subdividing the code into further levels is allowed, they'll get
aggregated into the allowed SAT code (of the form XXX or XXX.YY).
Dots in account names shouldn't matter.

This fix makes the parsing of the NumCta value more strict so it
supports the above use cases.

In case the prefix doesn't match the expected pattern, an error is
shown to the user. Normally this shouldn't happen, since the line
would have been filtered out if it didn't match a valid prefix from
the upper levels. It was added to make the code more robust against
refactoring and customization, and to prevent an uncaught traceback
in that case.

opw-3878763

closes odoo/enterprise#62920

X-original-commit: accfa4501b4957bd5ca537fafb3acc547e98c357
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Merel Geens <mege@odoo.com>