[IMP] pos_preparation_display: show instantly attribute in popup
Before this commit when a product have attribute in instantly mode, the
product was duplicated in the list of products as many times as there
were possible combinations.

Now, the product is displayed only once and when the user clicks on it,
a popup is displayed with the possible combinations.

closes odoo/enterprise#61907

Taskid: 3884515
Related: odoo/odoo#163781
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>