[FIX] account_batch_payment: make code readonly
payment_method_code was in the view for the batch payment creation view.
However, when you create batch payment the check printing module has a
compute that was being triggered on all payments with the same payment
method because the payment method's code was
being written by the payment_method_code on this view.

Setting this field to readonly causes the create to no longer use this
payment_method_code value and subsquently no longer triggers the
recompute.

opw-3848817

closes odoo/enterprise#62877

X-original-commit: 6091585b7330e03e429b746e964c256db8c173eb
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>