[IMP] sign: Prevent expiracy template when no timestamp is provided
Before this commit, when the sign_document_from_mail controller was
called without timestamp parameter, it would redirect to the
sign.sign_request_expired template because
current_request_item._validate_expiry(post.get('timestamp'),
post.get('exp')) would return False.

For human triggered interaction, it is not an issue.

Some customer complained of massive email sent. One of the signer is
probably relying on MS Defender. This is system is known to click on
links in email and cleaning the GET parameters of the URL.

In that case, the bot will get the sign.sign_request_expired which has a
button that allows to resend the sign request. If the bot click on that
button, it will send an email, that will be analyzed by the bot. it will
clean the parameters, get the sign.sign_request_expired template etc
etc.

opw-ticket 3848454

closes odoo/enterprise#61853

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>