[IMP] l10n_be_hr_payroll: include attestation days in holiday right computation.
When you use the holiday right wizard It computes,
based on the time you worked during the previous
year, how many official annual paid time off
you're eligible to.

When you switch company in the middle of the
year (let's say 01 June 2024), the previous
employer will give you an amount of money 
Holiday attest 2023 (Attestation N-1) :
it's the money for the days you have accumulated
from 01/01/2023 to 31/12/2023 and those days will
be used during the year 2024

Holiday attest 2024 (Attestation N) : it's the
money for the days you have accumulated
between 01/01/2024 to 31/05/2024 and that
you'll receive on the 1st January 2025

Previously, when you use the wizard to compute the
right to paid time off, it will count the time
that you spent at odoo in 2024
(from 01/06/2024 to 31/12/2024) and it won't take
into consideration the attestation days.

The provided change would include the attestation
days in the computation of paid time off right.

ALso, add 2 tests for allocating paid time
off with attestation days. In the first, we test
if the correct amount of paid time off is
allocated including the attestation days of year N.
In the next test, we check if we allocate
the correct amount of paid time off including
attestation days of year N-1.

task-3877999

closes odoo/enterprise#62050

Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>