[FIX] website_sale_subscription: handle recurring product with 1 plan
Steps
[website_sale_subscription]
- Settings > Website, set "Add to Cart" behaviour to "Go to cart"
- Create a recurring product with only one recurring price.
- Go to shop, choose the recurring product and try to add it to the cart
** Error 500 **

Issue
In case there's only one plan, no menu to select a plan is displayed and plan_id
is NaN in that case.
https://github.com/odoo/enterprise/blob/7d160198be779095660e7f6fb68ae231e4621972/website_sale_subscription/static/src/js/website_sale_subscription.js#L15
Which is going to be problematic later
https://github.com/odoo/enterprise/blob/7d160198be779095660e7f6fb68ae231e4621972/website_sale_subscription/models/sale_order.py#L21

introduced by 7d160198be779095660e7f6fb68ae231e4621972

Fix
We do not pass a plan_id if there's only one existing plan and use the fallback.

opw-3891366

closes odoo/enterprise#62641

X-original-commit: c540e48c6c3d469fb6e85659c0ecd705ce48bbf3
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>