[IMP] hr_payroll_account: Adapt payment register for multiple payslips
Currently, `action_register_payment` and the `account.payment.register`
wizard depends on the payslip in the context to only register payment
for one payslip. This restricts any way of registering payment for
a batch of payslip.

This commit, adapts the `hr_payroll_account` module to allow multiple
payslips to be registered as a batch. However, it maintains the current
flow for single payslip payment. This can be extended in other modules
to allow `hr.payslip.run` to register payments if needed.

Part-of: odoo/enterprise#62162