[FIX] knowledge: open article form view from thread notifications
This commit fixes an issue with the Knowledge Article Thread notifications where
the user would be redirected to the technical view instead of the Article's
view.

To fix this, the threadActionsRegistry has been updated for the `expand-form`
action so that if the thread model is a `knowledge.article.thread`, it will
redirect the user to the corresponding article.

task-3904795

closes odoo/enterprise#62024

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>