[FIX] sale_planning: auto plan open shifts without SOL set
Before this commit, the open shifts without any SOL set are not fetched
during the auto plan process because the following condition is falsy
for them inside the domain used: `('sale_line_id.state', '!=', False)`.

This commit changes the domain to check if the SOL is false or if the
state of the SOL set is different than 'cancel'.

runbot-61465

closes odoo/enterprise#62303

X-original-commit: bdf3f263dd069e7b1eb982468f381170244461ac
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>