[FIX] documents: traceback when navigating in split tool
Steps to reproduce:
1. Open the split tool, select a few pages, and create an expense for instance
2. Once the pages are gone, navigate with the keyboard
3: Traceback occurred

Technical reason:
When we navigated using the left arrow, it couldn't find the documents we were
targeting. In the '_focusNextPage' function, inside 'PdfManager' it was
attempting to access elements like 'arr[-2]', which resulted in an
'undefined' value.

After this commit:
No traceback will occur and if it can't find the previous or next selected
page, it focuses on the first page by default.

Task-3874151

closes odoo/enterprise#62244

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>