[FIX] web: decrease placeholder color
This commit aims to decrease a bit the contrast of placeholder text
in order to avoid confusion with regular text.

Prior to this commit, placeholder texts were using a gray value from the
grayscale, which was very similar to the `$body-color` in some
context.

To fix that issue, the color of the placeholder has been reduced to make
it lighter, even though it doesn't meet `3.5:1` contrast ratio.

task-3686510

closes odoo/enterprise#63150

X-original-commit: 7f9bf27afcfae046b289a5c7f3b330b80cc34010
Related: odoo/odoo#166652
Signed-off-by: Romeo Fragomeli (rfr) <rfr@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>