[IMP] hr_referral: referral campaign wizard and search views
This commit adds a wizard to create a referral campaign.

This wizard adds a new flow to launch a referral campaign from
hr_recruitment. It allows to send emails/sms to employees to invite them
to participate to a referral campaign. The campaign can be configured to
be open to all employees or only to employees of a specific department,
job position or just selected ones.
The purpose of this new flow is to give more light to the hr_referral
module and to allow to launch campaigns for specific jobs.

this commit also adds a new methods 'search_or_create_referral_links'
to hr.job model which will search for existing referral links
for the given job for a set of users or create new ones if
some are missing.

This method is used in the hr.referral.campaign.wizard to
generate referral links for the selected users. But this
method is so general that is prefered to have it in the
hr.job model.

And to finish this commit also change search view of job
in the referral app because the The original search was the same one
as the one used in the job search view from hr_recruitment.
This search was not adapted to the referral because
it was including filters that were already enforced by the referral
domain.

task-3607159

closes odoo/enterprise#51294

Related: odoo/odoo#172593
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>