[FIX] appointment_google_calendar: remove write override
This commit removes the write override of calendar.event records, since this override
was doing unnecessary requests on Google side when synchronizing again events that were
just synchronized.

After this commit, the function get_post_sync_values() will reuse the values returned by
the insertion in Google and write in the event right away, saving requests to Google. The
objective of the previous write override was to write the google meet URL in the event,
and now this is still being done correctly by means of the new strategy.

related-to: odoo/odoo#163364
task-389382

closes odoo/enterprise#61989

X-original-commit: 118817f12c7b6f731412819185ab37842332c1bb
Related: odoo/odoo#164422
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>