[FIX] hr_payroll: Make total not computed
Purpose
=======

Currently the amount, qty and rate are set on the payslip lines and the total is computed
afterwards. This works because the total is always equal to the amount * qty * rate.

But if some localisations want to alter the total amount with a specific round (like the 5 cents
rounding in switzerland), the recomputation is inefficient and could lead to issues because we
lack the computation context afterwards.

So, when creating the payslip lines, we should directly specify the total amount instead of
delegating the computation for later.

Part-of: odoo/enterprise#65174
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>