[FIX] documents_account: Allow selection of Bank/Cash journals
Before this commit, users were not allowed to activate the bridge to the documents app to work on Bank and Cash journals.
This commit allows Bank and Cash journal documents to be centralized the same way other journals are.

There does not seem to be particular reason about excluding Bank and Cash journals from the documents bridge.
The potential of showing unwanted bank account numbers in a general repository can always be managed by other means such as access rights.
The exclusion was added in this PR #10141 but it seems unnecessary and needs to be removed.

task-3839839

closes odoo/enterprise#62088

Signed-off-by: Quentin De Paoli <qdp@odoo.com>