[FIX] account_reports: Adapt UI for Analytic filters
In this commit [1], we forgot to have the same title for
2 filters: Analytic Filter and Analytic Group By.
The aim of this commit is having the same title for these 2
filters and distinct them by using correct icon.
We're using the group account from search view for the group by
and the filter one for the analytic filter.
In the mean time, we change the extra options one for something
meaning more that it's a setting than a filter.

no task id

[1]: https://github.com/odoo/enterprise/commit/a13a54921f8ae79e9ac3e2eb3f153a946a667192

closes odoo/enterprise#62112

X-original-commit: 696e48ec0690f75cd0c441ca95855a20307a216e
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>