[FIX] l10n_mx_edi_pos: use parter default CFDI usage in POS
Steps to reproduce:
1. In the contact page for a Mexican-based contact, add a default usage
2. Start a new session in point of sale
3. Select the client with a default usage, add go to payment
4. Select invoice. The usage field shows 'Acquisition of merchandise',
despit the default usage selected before

Currently, the popup only takes into account the l10n_mx_edi_usage
selected in the POS session or the code 'Acquisition of merchandise' if
the field is undefined.

The same behaviour happens if an invoice is created for a previous order.

This commit adds l10n_mx_edi_usage to the partner fields to be loaded when
loading the pos data. Then, the field in the popup can be filled with
the default CDFI usage of the partner if it exists. Else, it behaves
as before, displaying either the usage code selected in the POS session
or code 'Acquisition of merchandise' if the field is undefined.

This commit requires that the class InvoiceButton passes the selected
partner to the AddInfoPopup (odoo/odoo#166648).

task-3865331

closes odoo/enterprise#63147

X-original-commit: 1ac6ba2e6edf2cd6a97f5956b6002f13ebd8aad1
Signed-off-by: Laurent Smet (las) <las@odoo.com>