[FIX] voip: open record from correspondence details
Since the complete refactoring of the module in 17, the "open record"
button of the CorrespondenceDetails no longer works: Clicking it causes
the softphone to fold, but nothing more happens (the record is not
displayed as expected).

This is because we started retrieving the ORM service from the
useService hook, whereas before we used to retrieve it from the env.
useService untransparently wraps the ORM call in a "_protectMethod",
which drops the return value of an RPC if the component that made it has
been destroyed in the meantime. Since clicking the "open record" button
also folds the softphone, it destroyed the Correspondence Details
component, preventing it from ever receiving the RPC's return value.

This commit changes the way the ORM service is retrieved in order not to
use useService and to solve the problem.

opw-3904519

closes odoo/enterprise#62099

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>