[FIX] documents: Allow editing mail attachments from emails
When receiving emails, the attachments of the email may be made directly
into a document. This will not make a copy, meaning restrictions
on message attachments apply on these attachments. By extension
this means the related fields on documents cannot be used properly.

We now create a copy of the attachment with the replaced value when
modifying a document instead of making a copy and writing on the original.

This makes more sense globally and avoids write restrictions by just using
create and never writing datas to the attachments.

Relevant tests are updated with @users to cover non-system use cases.
Revision is improved and corrected. Test class is updated to setUpClass
as it's a cheap performance boost in tests.

issue introduced in odoo/odoo@4c4e63f01aefa01c2377f15f2020485cb0f7e4da

task-3519815

closes odoo/enterprise#62336

X-original-commit: 1b38ba9e7332347736c8d87dfff9a650a922b695
Related: odoo/odoo#165163
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>