[IMP] account_reports: Fixing column clipping on xlsx sheets
The way we calculate the column width is right, but we do it with the wrong font.
Currently, we use the metrics of the Lato font. This font is not available on customer computer, so we write on xlsx sheet with Arial instead. The difference between these two fonts can make metrics inaccurate and so adding a little extra padding to the column width ensure that the text will not be clipped.

Part-of: odoo/enterprise#59778