[FIX] voip: search contact/activity/call onInput
When typing in the softphone's search bar, the search result doesn't
appear immediately; the user has to press Enter to make it happen. This
is because the input field uses the t-model directive with the trim
modifier. The trim modifier silently switches the directive to lazy
mode, meaning that the content of the field is updated whenever the
onchange event is fired instead of oninput.

This commit removes the trim modifier to fix the problem.

opw-3890028

closes odoo/enterprise#62125

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>