[FIX] l10n_lu_reports: Properly check the enabled tax countries when deciding whether or not to show the Luxembourg-specific fields on an accounting firm
Before, only the fiscal country of the company was checked. Because of that, a Belgian company with a foreign VAT fiscal position in Luxembourg could for example not define those fields on its accounting firm, making it impossible for it to submit its tax declaration file to the government (since generating the file required those fields to be set).

opw-3875483

closes odoo/enterprise#62329

X-original-commit: 99aa5caa55a15037b0927c8871b2e3aa50dffc18
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>