[FIX] product_barcodelookup: neutralize
This commit adds the missing neutralization necessary for the product_barcodelookup
module introduced in [1]

The purpose of the standard neutralization framework is to allow us to
create database copies that will not interact with external systems in
ways that could impact the production database (or if it is not possible
to prevent the interactions, make sure that they are benign or wont
result in actual changes), or impact the customers of the operator of
the production database.

This is mainly useful to allow safe support investigation on database
duplicates.

[1] https://github.com/odoo/enterprise/pull/61949

closes odoo/enterprise#62221

X-original-commit: 93916010dbbd352916375c149aa67d05dd4f7120
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Alexandre Moens (mao) <mao@odoo.com>