[PERF] spreadsheet_edition: prefetch revision uuid
Since a055dfeb84ab, the field `current_revision_uuid` is in the documents
kanban view. Which means it can be computed for many documents.

The compute method loads all the revisions of spreadsheet documents.
Revisions can hold a lot of data (field `commands`), several hundreds
of Mb for large spreadsheets.

With this commit, we pre-fetch the only field that we need to prevent the ORM
to prefetch all fields that we don't need (and specifically avoid loading
the field `commands`)

closes odoo/enterprise#62502

Task: 3918564
X-original-commit: 87c2b0a0ea6a30ad7f5eed8ddbe1daa185b6a784
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>