[IMP] account_online_synchronization: journal-bank currency check
Previously, when you sync a bank journal with a bank account that have
different currency, odoo will import all the transaction with the amount
of the bank account, but in the currency of the journal.

For example: when importing 500 USD from the bank account to an IDR
journal, it will become 500 IDR.

This behavior also gets a bit weirder because the code checks the
currency of the bank with the company (which should be unrelated here).

This commit aims to fix that, and establish these new behaviors:

- The currency of the company is irrelevant. Only the journal matters.
  (because a journal can have different currency than the company's)
- If the journal currency has a mismatch with the bank account's:
  - If there's no existing entries in the journal, re-assign the
    journal's currency to the bank account's currency
  - If there's existing entries, prevent the import process from
    happening to prevent the mentioned weird behaviors.

closes odoo/enterprise#58820

Task-id: 3798134
Signed-off-by: Quentin De Paoli <qdp@odoo.com>