[FIX] knowledge: handle embedded views optional dependencies
Cohort views were patched to allow inserting them as an embed in a Knowledge
article, but `web_cohort` was not a dependency of `knowledge`.

It was an issue if the `web_cohort` (which is auto_install: True) module was
manually uninstalled, because it would result in a missing dependency (warning
in the browser console) preventing the insertion of any external view in
Knowledge.

The same issue was present for Gantt views (`web_gantt`).

Using an odoo runtime import, we can evaluate if the module was loaded and apply
the patch at that point. If the module was not loaded, the patch is not applied.

task-3918938

closes odoo/enterprise#63211

X-original-commit: a76c5f3c6fa63a16aff82e2132994cb8f85a110c
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>