[FIX] knowledge: handle embedded views optional dependencies
Cohort views were patched to allow inserting them as an embed in a Knowledge
article, but `web_cohort` was not a dependency of `knowledge`.

It was an issue if the `web_cohort` (which is auto_install: True) module was
manually uninstalled, because it would result in a missing dependency (warning
in the browser console) preventing the insertion of any external view in
Knowledge.

The same issue was present for Gantt views (`web_gantt`).

Using an odoo runtime import, we can evaluate if the module was loaded and apply
the patch at that point. If the module was not loaded, the patch is not applied.

Dependencies were added in [1] to allow HOOT tests to work, but this commit
will remove the need for that while still allowing those views to be inserted
in Knowledge.

[1]: https://github.com/odoo/enterprise/commit/f1326b20c819763026424d7d92a3d5489db60bc0

task-3918938

closes odoo/enterprise#63418

X-original-commit: 0816fee4abfab0d25e7a6f392ca32224831e85bf
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>