[IMP] timesheet_grid: merge To Validate menus in only one
Before this commit, since the merge of #53047, the current scale
chose by the user is now stored into the local storage, it means the `To
Validate > Last Week` and `To Validate > Last Month` can be merged into
one menu to display the `Last Period` and we will use the scale stored
into the local storage instead of forcing the scale in a menu.

This commit replaces `To Validate > Last week` and `To Validate > Last
Month` by `To Validate > Last Period` to take into account the scale
stored in the local storage and go back to the previous period.

task-3893907

Part-of: odoo/enterprise#62190