[IMP] web_grid,timesheet_grid: remove grid_range in context and dead code
Before this commit, since the local storage now store the range, it is
no longer needed to force to use another range by using `grid_range`
inside the context, this context was essentially used for the `To
Validate > Last Week` and `To Validate > Last Month` menus but those
menus will be replaced by `To Validate > Last Period`.

This commit removes the dead code because the group expand employee is
made in the JS code by doing an additional rpc inside the grid view for
`All timesheets` menu. This commit also removes the `grid_range` in the
context and the using of that context since it is no longer needed.

task-3893907

Part-of: odoo/enterprise#62190