[FIX] l10n_be_hr_payroll: fix complete month computation
In the BE loca, we were flexible on the notion of complete month.
This was useful for odoo but did not make sense for the localization.
A complete month of work is if you work from
the first work day of the month to the last.
Non-working days and public holidays are excluded.

closes odoo/enterprise#62489

Task: 3908658
X-original-commit: e34436b8d76452b0b0c89eb8c18e28abaac3c1ec
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>