[FIX] l10n_be_hr_payroll: fix complete month computation
In the BE loca, we were flexible on the notion of complete month.
This was useful for odoo but did not make sense for the localization.
A complete month of work is if you work from
the first work day of the month to the last.
Non-working days and public holidays are excluded.

closes odoo/enterprise#62258

Task: 3908658
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>